+++ map.h
@@ -9,7 +9,7 @@
 
 typedef struct {
 	Tile **grid; // First dim is y, second dim is x
-	unsigned int yLength, xLength;
+	unsigned short yLength, xLength;
 	Layer *mapLayer; // the map layer
 	Layer *rangeLayer; // the range layer
 } Map;
@@ -24,7 +24,7 @@ void init_move_grids(Team *team, Map *map);
 void free_move_grid(Team *team, Map *map);
 
 void find_range(Team *team, Map *map);
-void flood_fill(int y, int x, int move, Unit *unit, Map *map);
+void flood_fill(short y, short x, int move, Unit *unit, Map *map);
 
 void draw_range(Unit *unit, Map *map);
 void undraw_range(Unit *unit, Map *map);
